requirements.txt
discord.py[voice]
pytube
python-dotenv
beautifulsoup4
